32 bit
git-svn-id: https://svn.sosy-lab.org/software/sv-benchmarks/trunk@36 770dfece-24ca-4306-9060-4abab334d2a5

+++ ldv-regression/alt_test.c-unsafe.cil.c
@@ -35,7 +35,7 @@ int main(int argc , char **argv )
   {
   {
 #line 13
-  __cil_tmp5 = (int )4UL;
+  __cil_tmp5 = (int )4U;
 #line 13
   tmp = l_malloc(__cil_tmp5);
 #line 13
@@ -59,8 +59,8 @@ void *l_malloc(int size )
   void *tmp ;
   unsigned long __cil_tmp4 ;
   void *__cil_tmp5 ;
-  unsigned long __cil_tmp6 ;
-  unsigned long __cil_tmp7 ;
+  unsigned int __cil_tmp6 ;
+  unsigned int __cil_tmp7 ;
 
   {
   {
@@ -75,9 +75,9 @@ void *l_malloc(int size )
 #line 21
   __cil_tmp5 = (void *)0;
 #line 21
-  __cil_tmp6 = (unsigned long )__cil_tmp5;
+  __cil_tmp6 = (unsigned int )__cil_tmp5;
 #line 21
-  __cil_tmp7 = (unsigned long )retVal;
+  __cil_tmp7 = (unsigned int )retVal;
 #line 21
   if (__cil_tmp7 != __cil_tmp6) {
 #line 22
@@ -93,17 +93,17 @@ void *l_malloc(int size )
 #line 26 "files/alt_test.c"
 void l_free(void *ptr ) 
 { void *__cil_tmp2 ;
-  unsigned long __cil_tmp3 ;
-  unsigned long __cil_tmp4 ;
+  unsigned int __cil_tmp3 ;
+  unsigned int __cil_tmp4 ;
 
   {
   {
 #line 30
   __cil_tmp2 = (void *)0;
 #line 30
-  __cil_tmp3 = (unsigned long )__cil_tmp2;
+  __cil_tmp3 = (unsigned int )__cil_tmp2;
 #line 30
-  __cil_tmp4 = (unsigned long )ptr;
+  __cil_tmp4 = (unsigned int )ptr;
 #line 30
   if (__cil_tmp4 != __cil_tmp3) {
 #line 30
